add userId functionality to /notes endpoint

+++ routes/notes.js
@@ -2,14 +2,19 @@
 
 const express = require('express');
 const router = express.Router();
+const passport = require('passport');
 
 const mongoose = require('mongoose');
 
 const Note = require('../models/note');
 
+// Protect endpoints using JWT Strategy
+router.use('/notes', passport.authenticate('jwt', { session: false, failWithError: true }));
+
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/notes', (req, res, next) => {
   const { searchTerm, folderId, tagId } = req.query;
+  const userId = req.user.id;
 
   let filter = {};
   /**
@@ -35,6 +40,10 @@ router.get('/notes', (req, res, next) => {
     sort = projection;
   }
 
+  // if (userId) {
+  //   filter.userId = userId;
+  // }
+
   // if querying by folder, then add to filter
   if (folderId) {
     filter.folderId = folderId;
@@ -58,6 +67,7 @@ router.get('/notes', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/notes/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -65,7 +75,7 @@ router.get('/notes/:id', (req, res, next) => {
     return next(err);
   }
 
-  Note.findById(id)
+  Note.findOne({ _id: id, userId })
     .select('title content created folderId tags')
     .populate('tags')
     .then(result => {
@@ -76,11 +86,24 @@ router.get('/notes/:id', (req, res, next) => {
       }
     })
     .catch(next);
+
+  // Note.findById(id)
+  //   .select('title content created folderId tags')
+  //   .populate('tags')
+  //   .then(result => {
+  //     if (result) {
+  //       res.json(result);
+  //     } else {
+  //       next();
+  //     }
+  //   })
+  //   .catch(next);
 });
 
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/notes', (req, res, next) => {
   const { title, content, folderId, tags } = req.body;
+  const userId = req.user.id;
 
   /***** Never trust users - validate input *****/
   if (!title) {
@@ -89,7 +112,7 @@ router.post('/notes', (req, res, next) => {
     return next(err);
   }
 
-  const newItem = { title, content, tags };
+  const newItem = { title, content, folderId, tags, userId };
 
   Note.create(newItem)
     .then(result => {
@@ -102,6 +125,7 @@ router.post('/notes', (req, res, next) => {
 router.put('/notes/:id', (req, res, next) => {
   const { id } = req.params;
   const { title, content, folderId, tags } = req.body;
+  const userId = req.user.id;
 
   /***** Never trust users - validate input *****/
   if (!title) {
@@ -116,7 +140,7 @@ router.put('/notes/:id', (req, res, next) => {
     return next(err);
   }
 
-  const updateItem = { title, content, tags };
+  const updateItem = { title, content, tags, userId };
   
   if (mongoose.Types.ObjectId.isValid(folderId)) {
     updateItem.folderId = folderId;
@@ -140,8 +164,9 @@ router.put('/notes/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/notes/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
-  Note.findByIdAndRemove(id)
+  Note.findByIdAndRemove({ id, userId })
     .then(count => {
       if (count) {
         res.status(204).end();
